Refactored files, removed main func, added tests.
